uppgift 1 a-d

+++ src/Uppgift1/Figure.java
@@ -1,6 +1,6 @@
 package Uppgift1;
 
 public interface Figure {
-    double getArea();
-    double getCircumference();
+    int getArea();
+    int getCircumference();
 }
